+++ src/main/kotlin/models/BankAccount.kt
@@ -5,7 +5,7 @@ typealias AccountNumber = Int
 data class AccountDetails(
     val accountNumber: AccountNumber,
     val userName: UserName,
-    var balance: Int,
+    var balance: Double,
 )
